修复：合并逻辑修改之后的bug

+++ src/SpiderTool/SpiderUtility.cs
@@ -162,21 +162,41 @@ namespace SpiderTool
 
             var dirs = rootDirInfo.GetDirectories();
             var dirsCount = dirs.Count();
-            foreach (var currentDirInfo in dirs)
+
+            if (dirsCount > 0)
             {
-                cancellationToken.ThrowIfCancellationRequested();
-
-                var allFiles = currentDirInfo.GetFiles().OrderBy(x => x.CreationTime).ToList();
-                if (allFiles.Count == 0)
+                foreach (var currentDirInfo in dirs)
                 {
-                    Directory.Delete(currentDirInfo.FullName);
-                    continue;
-                }
+                    cancellationToken.ThrowIfCancellationRequested();
 
+                    var allFiles = currentDirInfo.GetFiles().OrderBy(x => x.CreationTime).ToList();
+                    if (allFiles.Count == 0)
+                    {
+                        Directory.Delete(currentDirInfo.FullName);
+                        continue;
+                    }
+
+                    var files = allFiles.Where(x => x.Extension.ToLower() == ".txt").OrderBy(x => x.CreationTime).Select(x => x.FullName).ToList();
+                    if (files.Count != 0)
+                    {
+                        var fileName = GetDirToName(rootDirInfo, currentDirInfo) + ".txt";
+                        var filePath = Path.Combine(rootDir, fileName);
+                        foreach (var file in files)
+                        {
+                            var txt = await File.ReadAllTextAsync(file, cancellationToken);
+                            await File.AppendAllTextAsync(filePath, txt.TrimEnd(), cancellationToken);
+                            File.Delete(file);
+                        }
+                    }
+                }
+            }
+            else 
+            {
+                var allFiles = rootDirInfo.GetFiles().OrderBy(x => x.CreationTime).ToList();
                 var files = allFiles.Where(x => x.Extension.ToLower() == ".txt").OrderBy(x => x.CreationTime).Select(x => x.FullName).ToList();
                 if (files.Count != 0)
                 {
-                    var fileName = GetDirToName(rootDirInfo, currentDirInfo) + ".txt";
+                    var fileName = rootDirInfo.Name + ".txt";
                     var filePath = Path.Combine(rootDir, fileName);
                     foreach (var file in files)
                     {
@@ -186,6 +206,7 @@ namespace SpiderTool
                     }
                 }
             }
+
         }
 
         private static string GetDirToName(DirectoryInfo rootDirInfo, DirectoryInfo currentDirInfo)
